#473 Split the user queries into two files - subscription based queries and 'other' queries. Added the base for user removal command.
