+++ Whoaverse/Whoaverse/Commands/User/Removal.cs
@@ -1,0 +1,6 @@
+﻿namespace Voat.Commands.User
+{
+    public static class Removal
+    {
+    }
+}
